requirements.txt
numpy>=1.24.0
pandas>=2.0.0
lightgbm>=4.0.0
shap>=0.42.0
scikit-learn>=1.3.0
joblib>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
